test_path_writer move to a file from helper
