+++ launchable/commands/split_subset.py
@@ -1,11 +1,10 @@
 import click
 import os
 
-
 from ..utils.env_keys import REPORT_ERROR_KEY
 from ..utils.http_client import LaunchableClient
 from ..utils.click import FRACTION
-from .subset import TestPathWriter
+from .test_path_writer import TestPathWriter
 
 
 @click.group(help="Split subsetting tests")
